new_scraper/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
frozenlist==1.3.3
google-search-results==2.4.1
idna==3.4
langchain==0.0.102
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.0
packaging==23.0
pycryptodomex==3.17
pydantic==1.10.5
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
SQLAlchemy==1.4.46
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.14
websockets==10.4
yarl==1.8.2
yt-dlp==2023.3.3
